Add SessionStore trait, which AtpAgent uses
