+++ atrium-api/src/agent.rs
@@ -1,5 +1,8 @@
 //! An ATP "Agent".
 //! Manages session token lifecycles and provides all XRPC methods.
+pub mod store;
+
+use self::store::SessionStore;
 use crate::client::Service;
 use async_trait::async_trait;
 use atrium_xrpc::error::{Error, XrpcErrorKind};
@@ -7,42 +10,43 @@ use atrium_xrpc::{HttpClient, OutputDataOrBytes, XrpcClient, XrpcRequest, XrpcRe
 use http::{Method, Request, Response};
 use serde::{de::DeserializeOwned, Serialize};
 use std::sync::Arc;
-use tokio::sync::{Mutex, Notify, RwLock};
+use tokio::sync::{Mutex, Notify};
 
 /// Type alias for the [com::atproto::server::create_session::Output](crate::com::atproto::server::create_session::Output)
 pub type Session = crate::com::atproto::server::create_session::Output;
 
-pub struct SessionAuthWrapper<T>
-where
-    T: XrpcClient + Send + Sync,
-{
-    session: Arc<RwLock<Option<Session>>>,
-    inner: T,
+const REFRESH_SESSION: &str = "com.atproto.server.refreshSession";
+
+pub struct SessionAuthWrapper<S, T> {
+    store: Arc<S>,
+    inner: Arc<T>,
 }
 
 #[async_trait]
-impl<T> HttpClient for SessionAuthWrapper<T>
+impl<S, T> HttpClient for SessionAuthWrapper<S, T>
 where
-    T: XrpcClient + Send + Sync,
+    S: Send + Sync,
+    T: HttpClient + Send + Sync,
 {
     async fn send_http(
         &self,
-        req: Request<Vec<u8>>,
+        request: Request<Vec<u8>>,
     ) -> Result<Response<Vec<u8>>, Box<dyn std::error::Error + Send + Sync + 'static>> {
-        self.inner.send_http(req).await
+        self.inner.send_http(request).await
     }
 }
 
 #[async_trait]
-impl<T> XrpcClient for SessionAuthWrapper<T>
+impl<S, T> XrpcClient for SessionAuthWrapper<S, T>
 where
+    S: SessionStore + Send + Sync,
     T: XrpcClient + Send + Sync,
 {
     fn base_uri(&self) -> &str {
         self.inner.base_uri()
     }
     async fn auth(&self, is_refresh: bool) -> Option<String> {
-        self.session.read().await.as_ref().map(|session| {
+        self.store.get_session().await.map(|session| {
             if is_refresh {
                 session.refresh_jwt.clone()
             } else {
@@ -52,18 +56,16 @@ where
     }
 }
 
-pub struct RefreshWrapper<T>
-where
-    T: XrpcClient + Send + Sync,
-{
-    session: Arc<RwLock<Option<Session>>>,
-    inner: T,
+pub struct RefreshWrapper<S, T> {
+    store: Arc<S>,
+    inner: Arc<T>,
     is_refreshing: Arc<Mutex<bool>>,
     notify: Arc<Notify>,
 }
 
-impl<T> RefreshWrapper<T>
+impl<S, T> RefreshWrapper<S, T>
 where
+    S: SessionStore,
     T: XrpcClient + Send + Sync,
 {
     // Internal helper to refresh sessions
@@ -84,21 +86,16 @@ where
     }
     async fn refresh_session_inner(&self) {
         if let Ok(output) = self.call_refresh_session().await {
-            let mut session = self.session.write().await;
-            let did_doc = session.as_ref().and_then(|s| s.did_doc.clone());
-            let email = session.as_ref().and_then(|s| s.email.clone());
-            let email_confirmed = session.as_ref().and_then(|s| s.email_confirmed);
-            session.replace(Session {
-                access_jwt: output.access_jwt,
-                did: output.did,
-                did_doc,
-                email,
-                email_confirmed,
-                handle: output.handle,
-                refresh_jwt: output.refresh_jwt,
-            });
+            if let Some(mut session) = self.store.get_session().await {
+                session.access_jwt = output.access_jwt;
+                session.did = output.did;
+                session.did_doc = output.did_doc;
+                session.handle = output.handle;
+                session.refresh_jwt = output.refresh_jwt;
+                self.store.set_session(session).await;
+            }
         } else {
-            self.session.write().await.take();
+            self.store.clear_session().await;
         }
     }
     // same as `crate::client::com::atproto::server::Service::refresh_session()`
@@ -112,7 +109,7 @@ where
             .inner
             .send_xrpc::<(), (), _, _>(&XrpcRequest {
                 method: Method::POST,
-                path: "com.atproto.server.refreshSession".into(),
+                path: REFRESH_SESSION.into(),
                 parameters: None,
                 input: None,
                 encoding: None,
@@ -140,21 +137,23 @@ where
 }
 
 #[async_trait]
-impl<T> HttpClient for RefreshWrapper<T>
+impl<S, T> HttpClient for RefreshWrapper<S, T>
 where
-    T: XrpcClient + Send + Sync,
+    S: Send + Sync,
+    T: HttpClient + Send + Sync,
 {
     async fn send_http(
         &self,
-        req: Request<Vec<u8>>,
+        request: Request<Vec<u8>>,
     ) -> Result<Response<Vec<u8>>, Box<dyn std::error::Error + Send + Sync + 'static>> {
-        self.inner.send_http(req).await
+        self.inner.send_http(request).await
     }
 }
 
 #[async_trait]
-impl<T> XrpcClient for RefreshWrapper<T>
+impl<S, T> XrpcClient for RefreshWrapper<S, T>
 where
+    S: SessionStore + Send + Sync,
     T: XrpcClient + Send + Sync,
 {
     fn base_uri(&self) -> &str {
@@ -181,33 +180,37 @@ where
     }
 }
 
-pub struct AtpAgent<T>
+type Api<S, T> = Service<RefreshWrapper<S, SessionAuthWrapper<S, T>>>;
+
+pub struct AtpAgent<S, T>
 where
+    S: SessionStore + Send + Sync,
     T: XrpcClient + Send + Sync,
 {
-    pub api: Service<RefreshWrapper<SessionAuthWrapper<T>>>,
-    session: Arc<RwLock<Option<Session>>>,
+    store: Arc<S>,
+    pub api: Api<S, T>,
 }
 
-impl<T> AtpAgent<T>
+impl<S, T> AtpAgent<S, T>
 where
+    S: SessionStore + Send + Sync,
     T: XrpcClient + Send + Sync,
 {
-    pub fn new(xrpc: T) -> Self {
-        let session = Arc::new(RwLock::new(None));
+    pub fn new(xrpc: T, store: S) -> Self {
+        let store = Arc::new(store);
         let api = Service::new(Arc::new(RefreshWrapper {
-            session: Arc::clone(&session),
-            inner: SessionAuthWrapper {
-                session: Arc::clone(&session),
-                inner: xrpc,
-            },
+            store: Arc::clone(&store),
+            inner: Arc::new(SessionAuthWrapper {
+                store: Arc::clone(&store),
+                inner: Arc::new(xrpc),
+            }),
             is_refreshing: Arc::new(Mutex::new(false)),
             notify: Arc::new(Notify::new()),
         }));
-        Self { api, session }
+        Self { api, store }
     }
     pub async fn get_session(&self) -> Option<Session> {
-        self.session.read().await.clone()
+        self.store.get_session().await
     }
     /// Start a new session with this agent.
     pub async fn login(
@@ -225,7 +228,7 @@ where
                 password: password.into(),
             })
             .await?;
-        self.session.write().await.replace(result.clone());
+        self.store.set_session(result.clone()).await;
         Ok(result)
     }
     /// Resume a pre-existing session with this agent.
@@ -233,20 +236,22 @@ where
         &self,
         session: Session,
     ) -> Result<(), Error<crate::com::atproto::server::get_session::Error>> {
-        self.session.write().await.replace(session.clone());
+        self.store.set_session(session.clone()).await;
         let result = self.api.com.atproto.server.get_session().await;
         match result {
             Ok(output) => {
                 assert_eq!(output.did, session.did);
-                if let Some(session) = self.session.write().await.as_mut() {
+                if let Some(mut session) = self.get_session().await {
+                    session.did_doc = output.did_doc;
                     session.email = output.email;
                     session.email_confirmed = output.email_confirmed;
                     session.handle = output.handle;
+                    self.store.set_session(session).await;
                 }
                 Ok(())
             }
             Err(err) => {
-                self.session.write().await.take();
+                self.store.clear_session().await;
                 Err(err)
             }
         }
@@ -254,334 +259,4 @@ where
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use futures::future::join_all;
-    use std::collections::HashMap;
-
-    #[derive(Default)]
-    struct DummyResponses {
-        create_session: Option<crate::com::atproto::server::create_session::Output>,
-        get_session: Option<crate::com::atproto::server::get_session::Output>,
-    }
-
-    #[derive(Default)]
-    struct DummyClient {
-        responses: DummyResponses,
-        counts: Arc<RwLock<HashMap<String, usize>>>,
-    }
-
-    #[async_trait]
-    impl HttpClient for DummyClient {
-        async fn send_http(
-            &self,
-            request: Request<Vec<u8>>,
-        ) -> Result<Response<Vec<u8>>, Box<dyn std::error::Error + Send + Sync + 'static>> {
-            tokio::time::sleep(std::time::Duration::from_micros(10)).await;
-            let builder =
-                Response::builder().header(http::header::CONTENT_TYPE, "application/json");
-            let token = request
-                .headers()
-                .get(http::header::AUTHORIZATION)
-                .and_then(|value| value.to_str().ok())
-                .and_then(|value| value.split(' ').last());
-            if token == Some("expired") {
-                return Ok(builder.status(http::StatusCode::BAD_REQUEST).body(
-                    serde_json::to_vec(&atrium_xrpc::error::ErrorResponseBody {
-                        error: Some(String::from("ExpiredToken")),
-                        message: Some(String::from("Token has expired")),
-                    })?,
-                )?);
-            }
-            let mut body = Vec::new();
-            if let Some(nsid) = request.uri().path().strip_prefix("/xrpc/") {
-                *self.counts.write().await.entry(nsid.into()).or_default() += 1;
-                match nsid {
-                    "com.atproto.server.createSession" => {
-                        if let Some(output) = &self.responses.create_session {
-                            body.extend(serde_json::to_vec(output)?);
-                        }
-                    }
-                    "com.atproto.server.getSession" => {
-                        if token == Some("access") {
-                            if let Some(output) = &self.responses.get_session {
-                                body.extend(serde_json::to_vec(output)?);
-                            }
-                        }
-                    }
-                    "com.atproto.server.refreshSession" => {
-                        if token == Some("refresh") {
-                            body.extend(serde_json::to_vec(
-                                &crate::com::atproto::server::refresh_session::Output {
-                                    access_jwt: String::from("access"),
-                                    did: String::from("did"),
-                                    did_doc: None,
-                                    handle: String::from("handle"),
-                                    refresh_jwt: String::from("refresh"),
-                                },
-                            )?);
-                        }
-                    }
-                    _ => {}
-                }
-            }
-            if body.is_empty() {
-                Ok(builder
-                    .status(http::StatusCode::UNAUTHORIZED)
-                    .body(serde_json::to_vec(
-                        &atrium_xrpc::error::ErrorResponseBody {
-                            error: Some(String::from("AuthenticationRequired")),
-                            message: Some(String::from("Invalid identifier or password")),
-                        },
-                    )?)?)
-            } else {
-                Ok(builder.status(http::StatusCode::OK).body(body)?)
-            }
-        }
-    }
-
-    impl XrpcClient for DummyClient {
-        fn base_uri(&self) -> &str {
-            "http://localhost:8080"
-        }
-    }
-
-    fn session() -> Session {
-        Session {
-            access_jwt: String::from("access"),
-            did: String::from("did"),
-            did_doc: None,
-            email: None,
-            email_confirmed: None,
-            handle: String::from("handle"),
-            refresh_jwt: String::from("refresh"),
-        }
-    }
-
-    #[tokio::test]
-    async fn test_new() {
-        let agent = AtpAgent::new(DummyClient::default());
-        assert_eq!(agent.get_session().await, None);
-    }
-
-    #[tokio::test]
-    async fn test_login() {
-        let session = session();
-        // success
-        {
-            let client = DummyClient {
-                responses: DummyResponses {
-                    create_session: Some(crate::com::atproto::server::create_session::Output {
-                        ..session.clone()
-                    }),
-                    ..Default::default()
-                },
-                ..Default::default()
-            };
-            let agent = AtpAgent::new(client);
-            agent
-                .login("test", "pass")
-                .await
-                .expect("login should be succeeded");
-            assert_eq!(agent.get_session().await, Some(session));
-        }
-        // failure with `createSession` error
-        {
-            let client = DummyClient {
-                responses: DummyResponses {
-                    ..Default::default()
-                },
-                ..Default::default()
-            };
-            let agent = AtpAgent::new(client);
-            agent
-                .login("test", "bad")
-                .await
-                .expect_err("login should be failed");
-            assert_eq!(agent.get_session().await, None);
-        }
-    }
-
-    #[tokio::test]
-    async fn test_xrpc_get_session() {
-        let session = session();
-        let client = DummyClient {
-            responses: DummyResponses {
-                get_session: Some(crate::com::atproto::server::get_session::Output {
-                    did: session.did.clone(),
-                    did_doc: session.did_doc.clone(),
-                    email: session.email.clone(),
-                    email_confirmed: session.email_confirmed,
-                    handle: session.handle.clone(),
-                }),
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-        let agent = AtpAgent::new(client);
-        agent.session.write().await.replace(session);
-        let output = agent
-            .api
-            .com
-            .atproto
-            .server
-            .get_session()
-            .await
-            .expect("get session should be succeeded");
-        assert_eq!(output.did, "did");
-    }
-
-    #[tokio::test]
-    async fn test_xrpc_get_session_with_refresh() {
-        let mut session = session();
-        session.access_jwt = String::from("expired");
-        let client = DummyClient {
-            responses: DummyResponses {
-                get_session: Some(crate::com::atproto::server::get_session::Output {
-                    did: session.did.clone(),
-                    did_doc: session.did_doc.clone(),
-                    email: session.email.clone(),
-                    email_confirmed: session.email_confirmed,
-                    handle: session.handle.clone(),
-                }),
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-        let agent = AtpAgent::new(client);
-        agent.session.write().await.replace(session);
-        let output = agent
-            .api
-            .com
-            .atproto
-            .server
-            .get_session()
-            .await
-            .expect("get session should be succeeded");
-        assert_eq!(output.did, "did");
-        assert_eq!(
-            agent.get_session().await.map(|session| session.access_jwt),
-            Some("access".into())
-        );
-    }
-
-    #[tokio::test]
-    async fn test_xrpc_get_session_with_duplicated_refresh() {
-        let mut session = session();
-        session.access_jwt = String::from("expired");
-        let client = DummyClient {
-            responses: DummyResponses {
-                get_session: Some(crate::com::atproto::server::get_session::Output {
-                    did: session.did.clone(),
-                    did_doc: session.did_doc.clone(),
-                    email: session.email.clone(),
-                    email_confirmed: session.email_confirmed,
-                    handle: session.handle.clone(),
-                }),
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-        let counts = Arc::clone(&client.counts);
-        let agent = Arc::new(AtpAgent::new(client));
-        agent.session.write().await.replace(session);
-        let handles = (0..3).map(|_| {
-            let agent = Arc::clone(&agent);
-            tokio::spawn(async move { agent.api.com.atproto.server.get_session().await })
-        });
-        let results = join_all(handles).await;
-        for result in &results {
-            let output = result
-                .as_ref()
-                .expect("task should be successfully executed")
-                .as_ref()
-                .expect("get session should be succeeded");
-            assert_eq!(output.did, "did");
-        }
-        assert_eq!(
-            agent.get_session().await.map(|session| session.access_jwt),
-            Some("access".into())
-        );
-        assert_eq!(
-            counts.read().await.clone(),
-            HashMap::from_iter([
-                ("com.atproto.server.refreshSession".into(), 1),
-                ("com.atproto.server.getSession".into(), 3)
-            ])
-        );
-    }
-
-    #[tokio::test]
-    async fn test_resume_session() {
-        let session = session();
-        // success
-        {
-            let client = DummyClient {
-                responses: DummyResponses {
-                    get_session: Some(crate::com::atproto::server::get_session::Output {
-                        did: session.did.clone(),
-                        did_doc: session.did_doc.clone(),
-                        email: session.email.clone(),
-                        email_confirmed: session.email_confirmed,
-                        handle: session.handle.clone(),
-                    }),
-                    ..Default::default()
-                },
-                ..Default::default()
-            };
-            let agent = AtpAgent::new(client);
-            assert_eq!(agent.get_session().await, None);
-            agent
-                .resume_session(Session {
-                    email: Some(String::from("test@example.com")),
-                    ..session.clone()
-                })
-                .await
-                .expect("resume_session should be succeeded");
-            assert_eq!(agent.get_session().await, Some(session.clone()));
-        }
-        // failure with `getSession` error
-        {
-            let client = DummyClient {
-                responses: DummyResponses {
-                    ..Default::default()
-                },
-                ..Default::default()
-            };
-            let agent = AtpAgent::new(client);
-            assert_eq!(agent.get_session().await, None);
-            agent
-                .resume_session(session)
-                .await
-                .expect_err("resume_session should be failed");
-            assert_eq!(agent.get_session().await, None);
-        }
-    }
-
-    #[tokio::test]
-    async fn test_resume_session_with_refresh() {
-        let session = session();
-        let client = DummyClient {
-            responses: DummyResponses {
-                get_session: Some(crate::com::atproto::server::get_session::Output {
-                    did: session.did.clone(),
-                    did_doc: session.did_doc.clone(),
-                    email: session.email.clone(),
-                    email_confirmed: session.email_confirmed,
-                    handle: session.handle.clone(),
-                }),
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-        let agent = AtpAgent::new(client);
-        agent
-            .resume_session(Session {
-                access_jwt: "expired".into(),
-                ..session.clone()
-            })
-            .await
-            .expect("resume_session should be succeeded");
-        assert_eq!(agent.get_session().await, Some(session));
-    }
-}
+mod tests;
